requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.14.0
idna==3.10
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2024.2
requests==2.28.1
smsir-python==1.0.8
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
